requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.1
annotated-types==0.7.0
asttokens==3.0.0
autoflake==2.3.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
greenlet==3.2.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
noise==1.2.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.90
optuna==4.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.3.2
Pygments==2.19.1
pykwalify==1.8.0
pyparsing==3.2.3
pyradiomics==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
simpleitk==2.5.0
simsimd==6.5.12
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.144
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
